# apps/partner-runtime/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CredWorks Education partner runtime",
    "version": "0.1.0",
    "description": "Supported profile: W3C SD-JWT+KB, ES256, DCQL literal scalar Education paths, decentralized_identifier request client ID, direct_post vp_token. Ephemeral sessions/cache, persistent encrypted Rust identity. Container/public distribution and physical observations are separately recorded. Stable codes and limitations are documented in README."
  },
  "paths": {
...
    "/management/sessions": {
      "post": {
        "operationId": "createSession",
        "servers": [
          {
            "url": "http://127.0.0.1:3081",
            "description": "Protected application listener in runtime network namespace"
          }
        ],
        "security": [
          {
            "RuntimeBearer": []
          }
        ],
        "description": "Requires fresh exact issuer/type/definition/profile/origin/key scope. Protect returned correlation_capability in the initiating application interaction. No account, eligibility or cookie decision. Maximum100 retained sessions; management bodies4KiB.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSession"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Stable refusal; no diagnostics or claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Stable refusal; no diagnostics or claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "408": {
            "description": "Stable refusal; no diagnostics or claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Stable refusal; no diagnostics or claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Stable refusal; no diagnostics or claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Stable refusal; no diagnostics or claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Runtime bearer refusal or protected correlation refusal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "RuntimeBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Runtime management token; registry project credential does not authorize runtime"
      },
      "SessionCapability": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Capability",
        "description": "Protected application interaction capability returned on creation; distinct from public state/request capability"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "description": "Runtime codes: SESSION_BAD_REQUEST, RESPONSE_BAD_REQUEST, RESPONSE_STATE_MISMATCH, SESSION_ACCESS_DENIED, SCOPE_NOT_PERMITTED, SESSION_NOT_FOUND, REQUEST_DEADLINE, REQUEST_TOO_LARGE, SESSION_CAPACITY, RESULT_PENDING, RESULT_CONSUMED, SESSION_ALREADY_COMPLETED, SESSION_EXPIRED, EVIDENCE_UNAVAILABLE, EVIDENCE_STALE, EVIDENCE_REFRESH_FAILED, PERMISSION_VERSION_UNSUPPORTED. Protected refused results additionally preserve validated identity-core enum codes, or PRESENTATION_VERIFICATION_FAILED for unknown exceptions; no diagnostic text."
              }
            },
            "required": [
              "code"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "CreateSession": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "profile": {
            "type": "string",
            "enum": [
              "education_eligibility",
              "education_sign_in"
            ]
          },
          "interaction_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "purpose": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "profile",
          "interaction_id",
          "purpose"
        ]
      },
      "Session": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "session_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "interaction_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "correlation_capability": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "request_uri": {
            "type": "string",
            "format": "uri"
          },
          "activation_uri": {
            "type": "string",
            "format": "uri"
          },
          "expires_at": {
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "interaction_id",
          "correlation_capability",
          "request_uri",
          "activation_uri",
          "expires_at"
        ]
      },
...
      "UnauthorizedError": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "error"
            ],
            "properties": {
              "error": {
                "const": "unauthorized"
              }
            }
          }
        ]
      }
    }
  }
}